rename kmskey resource to encryptionkey
