+++ service/controller/v13/resource/encryptionkey/create_test.go
@@ -1,4 +1,4 @@
-package kmskey
+package encryptionkey
 
 import (
 	"context"
