+++ service/controller/v13/resource/encryptionkey/current_test.go
@@ -1,4 +1,4 @@
-package kmskey
+package encryptionkey
 
 import (
 	"context"
