+++ service/controller/v13/resource/encryptionkey/delete_test.go
@@ -1,4 +1,4 @@
-package kmskey
+package encryptionkey
 
 import (
 	"context"
