+++ service/controller/v13/resource/encryptionkey/desired.go
@@ -1,4 +1,4 @@
-package kmskey
+package encryptionkey
 
 import (
 	"context"
