+++ service/controller/v13/resource/encryptionkey/error.go
@@ -1,4 +1,4 @@
-package kmskey
+package encryptionkey
 
 import (
 	"github.com/aws/aws-sdk-go/aws/awserr"
