+++ service/controller/v13/resource/encryptionkey/resource.go
@@ -1,4 +1,4 @@
-package kmskey
+package encryptionkey
 
 import (
 	"fmt"
